@@ -32,26 +32,38 @@
     "@babel/cli": "^7.14.8",
     "@babel/core": "^7.15.0",
     "@babel/preset-env": "^7.15.0",
+    "@types/cors": "^2.8.12",
+    "@types/debug": "^4.1.7",
+    "@types/express": "^4.17.13",
+    "@types/node": "^16.7.10",
     "@typescript-eslint/eslint-plugin": "^4.29.2",
     "@typescript-eslint/parser": "^4.29.2",
     "eslint": "^7.32.0",
     "eslint-config-google": "^0.14.0",
     "husky": "^7.0.1",
     "prettier": "^2.3.2",
-    "typescript": "^4.3.5",
+    "source-map-support": "^0.5.19",
+    "ts-node": "^10.2.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.2",
     "webpack": "^5.51.2",
     "webpack-cli": "^4.8.0"
   },
   "dependencies": {
+    "cors": "^2.8.5",
+    "debug": "^4.3.2",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
     "express-handlebars": "^5.3.3",
+    "express-winston": "^4.2.0",
     "pug": "^3.0.2",
     "rxjs": "^7.3.0",
-    "socket.io": "^4.1.3"
+    "socket.io": "^4.1.3",
+    "winston": "^3.3.3"
   },
   "engines": {
     "node": ">=14.0.0",
     "npm": ">=6.0.0"
-  }
+  },
+  "type": "module"
 }
